Suppress an occasionally flaky, randomized test.
Suppress an occasionally flaky, randomized test. The test sometimes
"hangs", possibly due to b/6428842 where it is suggested that the
randomized geometry the test generates can be "inverted", i.e.
instead of a small area, the area described is everywhere except for
a small area, leading to a massive area.

This is a guess - the test is being suppressed on Android as this
code isn't being used for anything important right now.

Bug: 280346791
Test: treehugger only
Change-Id: Id5ba8250839ac44b1515305431400328a7683744